// src/Districtr/Districtr.types.ts
export interface UnitProps {
  /** The unit number */
  id: number
  /** The district name */
  name?: string
  /** The district type */
  type?: 'single' | 'multi' | 'multi-varying'
  /** Number of seats available. Not a sub unit.  */
  members?: number
  /** The district color */
  color?: string
  /** The district color when hovered. */
  hoverColor?: string
  /** The district color when selected. */
  selectedColor?: string
  /** The district color when locked. */
  lockedColor?: string
  /** The district color when disabled */
  disabledColor?: string
  /** The population of the unit. */
  population?: number
  /** The ideal population of the unit. */
  idealPopulation?: number
  unitIdealPopulation?: number
  columnPopulations?: any
  /** The total number of units. */
  totalUnits?: number
  /** Remove from stack */
  stashed?: boolean
  note?: string
}
// ...
export interface UnitConfigProps {
  [key: number]: UnitProps
}